requirements.txt
requests
pyyaml
tweepy